feat: moved router code from main.jsx to rounter/index.jsx
